% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cache_activate}
\alias{cache_clear}
\alias{cache_deactivate}
\alias{cache_info}
\alias{create_style_guide}
\alias{default_style_guide_attributes}
\alias{specify_math_token_spacing}
\alias{specify_reindention}
\alias{specify_transformers_drop}
\alias{tidyverse_math_token_spacing}
\alias{tidyverse_reindention}
\alias{tidyverse_style}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{styler}{\code{\link[styler]{cache_activate}}, \code{\link[styler]{cache_clear}}, \code{\link[styler:cache_activate]{cache_deactivate}}, \code{\link[styler]{cache_info}}, \code{\link[styler]{create_style_guide}}, \code{\link[styler]{default_style_guide_attributes}}, \code{\link[styler:math_token_spacing]{specify_math_token_spacing}}, \code{\link[styler:reindention]{specify_reindention}}, \code{\link[styler]{specify_transformers_drop}}, \code{\link[styler:math_token_spacing]{tidyverse_math_token_spacing}}, \code{\link[styler:reindention]{tidyverse_reindention}}, \code{\link[styler]{tidyverse_style}}}
}}

