% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAndStylerEqualsAddin.R
\name{saveAndStylerEqualsAddin}
\alias{saveAndStylerEqualsAddin}
\title{Save and styler equal style active doc}
\usage{
saveAndStylerEqualsAddin()
}
\description{
Save and styler equal style active doc
}
