% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export.R
\name{style_file}
\alias{style_file}
\title{Like \code{\link[styler:style_file]{styler::style_file()}}, but \code{style} defaulting to \code{equals_style}}
\usage{
style_file(
  path,
  ...,
  style = equals_style,
  transformers = style(...),
  include_roxygen_examples = TRUE,
  base_indention = 0,
  dry = "off"
)
}
\arguments{
\item{path}{A character vector with paths to files to style. Supported
extensions: \code{.R}, \code{.Rmd}, \code{.Rmarkdown}, \code{.qmd} and \code{.Rnw}.}

\item{...}{Arguments passed on to the \code{style} function,
see \code{\link[styler:tidyverse_style]{tidyverse_style()}} for the default argument.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[styler]{tidyverse_style}}. Not used
further except to construct the argument \code{transformers}. See
\code{\link[styler:style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}

\item{include_roxygen_examples}{Whether or not to style code in roxygen
examples.}

\item{base_indention}{Integer scalar indicating by how many spaces the whole
output text should be indented. Note that this is not the same as splitting
by line and add a \code{base_indention} spaces before the code in the case
multi-line strings are present. See 'Examples'.}

\item{dry}{To indicate whether styler should run in \emph{dry} mode, i.e. refrain
from writing back to files .\code{"on"} and \code{"fail"} both don't write back, the
latter returns an error if the input code is not identical to the result
of styling. "off", the default, writes back if the input and output of
styling are not identical.}
}
\description{
See \code{\link[styler:style_file]{styler::style_file()}} for details, examples and more.
}
