% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export.R
\name{style_text}
\alias{style_text}
\title{Like \code{\link[styler:style_text]{styler::style_text()}}, but \code{style} defaulting to \code{equals_style}}
\usage{
style_text(
  text,
  ...,
  style = equals_style,
  transformers = style(...),
  include_roxygen_examples = TRUE,
  base_indention = 0
)
}
\arguments{
\item{text}{A character vector with text to style.}

\item{...}{Arguments passed on to the \code{style} function,
see \code{\link[styler:tidyverse_style]{tidyverse_style()}} for the default argument.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[styler]{tidyverse_style}}. Not used
further except to construct the argument \code{transformers}. See
\code{\link[styler:style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}

\item{include_roxygen_examples}{Whether or not to style code in roxygen
examples.}

\item{base_indention}{Integer scalar indicating by how many spaces the whole
output text should be indented. Note that this is not the same as splitting
by line and add a \code{base_indention} spaces before the code in the case
multi-line strings are present. See 'Examples'.}
}
\description{
See \code{\link[styler:style_text]{styler::style_text()}} for details, examples and more.
}
