% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-line-break.R
\name{force_assignment_eq}
\alias{force_assignment_eq}
\title{Force the assignment operator \code{=} when possible}
\usage{
force_assignment_eq(pd)
}
\description{
Because you can't determine from a nest with \verb{<-} what the context is,
we need to do it from the parent.
}
\keyword{internal}
